\name{basisplot}
\alias{basisplot}
\title{Generic basis plot function}
\usage{
basisplot(x, \dots)
}
\arguments{
\item{x}{basis to plot}
\item{\dots}{various arguments to methods}
}
\description{
Plots a representation of a time-frequency plane and then plots the locations,
and sometimes time series representations of coefficients, for the packets
in the basis.
}
\details{
Description says all
}
\value{
Nothing, usually
}
\seealso{\code{\link{basisplot.BP}}, \code{\link{basisplot.wp}}}
\author{G P Nason}
\keyword{hplot}
