\name{CanUseMoreThanOneColor}
\alias{CanUseMoreThanOneColor}
\title{Deprecated  function}
\usage{
CanUseMoreThanOneColor()
}
\description{
Not used any more. This function used to interrogate the display device
to see whether more than one color could be used. The function is set
to return true whether of not the display device actually has this capability.
It is used in the \code{\link{plot.wp}} function.
}
\details{
Description says it all.
}
\value{
This function always returns TRUE
}
\seealso{\code{\link{plot.wp}}}
\author{G P Nason}
\keyword{misc}
