\name{ConvertMessage}
\alias{ConvertMessage}
\title{Print out a text message about an object which is from old version of WaveThresh}
\usage{ConvertMessage()}
\arguments{None}
\description{
Print out text message about an object being from an old version of WaveThresh.
}
\details{
Description says all!
}
\value{
None
}
\seealso{\code{\link{IsEarly.default}},\code{\link{IsEarly}}}
\author{G P Nason}
\keyword{error}
