\name{drawbox}
\alias{drawbox}
\title{Draw a shaded coloured box}
\usage{
drawbox(x,y,w,h,density,col)
}
\arguments{
\item{x}{The bottom left x coordinate of the box}
\item{y}{The bottom left y coordinate of the box}
\item{w}{The width of the box}
\item{h}{The height of the box}
\item{density}{The shading density of the box}
\item{col}{The colour of the box}
}
\description{
Simply draws a box with bottom left corner at (x,y), or width w and height
h with shading of density and colour of col.
}
\details{
Description says all
}
\value{
None
}
\seealso{\code{\link{addpkt}}}
\author{G P Nason}
\keyword{dplot}
