\name{imwr}
\alias{imwr}
\title{2D Inverse Discrete Wavelet Transform (Image W. Reconstruction)}
\description{
  Generic function for 2-dimensional wavelet reconstruction.
}
\usage{
imwr(x, \dots)
}
\arguments{
\item{x}{an \R object.}
\item{...}{methods may have additional arguments.}
}
\seealso{the principal method, \code{\link{imwr.imwd},
    and \code{\link{imwr.imwdc}}.
}
\keyword{nonlinear}
\keyword{smooth}
% Converted by Sd2Rd version 0.3-3.
