\name{InvBasis}
\alias{InvBasis}
\title{Generic basis inversion for libraries}
\usage{
InvBasis(...)
}
\arguments{
\item{...}{Usually a library representation and a basis specification}
}
\description{
Will invert either a \code{wst} or \code{wp} object given that object
and some kind of basis specification.
}
\details{
Description says it all
}
\value{
The reconstruction.
}
\seealso{\code{\link{InvBasis.wp}},\code{\link{InvBasis.wst}},\code{\link{MaNoVe}},\code{\link{numtonv}}}
\author{G P Nason}
\keyword{smooth}
