\name{IsEarly}
\alias{IsEarly}
\title{Generic function to detect whether object is from an early version}
\usage{
IsEarly(x)
}
\arguments{
\item{x}{The object you want to see whether its from an early version}
}
\description{
Generic function to detect whether object is from an early version of
WaveThresh
}
\details{
Description says all
}
\value{
Returns TRUE if object is from an earlier version of WaveThresh, FALSE if
not.
}
\seealso{\code{\link{ConvertMessage}},\code{\link{IsEarly.default}},\code{\link{IsEarly}}, \code{\link{IsEarly.wd}}}
\author{G P Nason}
\keyword{error}
