\name{lennon}
\docType{data}
\alias{lennon}
\title{John Lennon image.}
\description{
A 256x256 matrix. Each entry of the matrix contains an image intensity value. The whole matrix represents an image of John Lennon
}
\usage{
data(lennon)
}
\format{
A 256x256 matrix. Each entry of the matrix contains an image intensity value. The whole matrix represents an image of John Lennon
}
\source{
The John Lennon image was supplied uncredited on certain UNIX workstations as an examples image. I am not sure who the Copyright belongs to. Please let me know if you know
}
\examples{
#
# This command produces the image seen above.
#
# image(lennon)
#
}
\keyword{datasets}
\author{G P Nason}
