\name{LocalSpec}
\alias{LocalSpec}
\title{Compute Nason and Silverman smoothed wavelet periodogram.}
\description{
This function is obsolete. Use the function \code{\link{ewspec}}. Performs the Nason and Silverman smoothed wavelet periodogram as described in Nason and Silverman (1995). 

This function is generic. 

Particular methods exist. For the wd class object this generic function uses \code{\link{LocalSpec.wd}}. 
}
\usage{
LocalSpec(...)
}
\arguments{
\item{\dots}{See individual help pages for details.}
}
\details{
See individual method help pages for operation and examples. 
}
\value{
The LocalSpec of the wavelet object supplied. See method help files for examples. 
}
\section{RELEASE}{Version 3.9 Copyright Guy Nason 1997 }
\seealso{
\code{\link{wd}}, \code{\link{wd.object}}, \code{\link{LocalSpec.wd}} 
}

\keyword{methods}
\author{G P Nason}
