\name{logabs}
\alias{logabs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Take the logarithm of the squares of the argument
}
\description{
Take the log of the squares of the argument
}
\usage{
logabs(x)
}
\arguments{
  \item{x}{A number
}
}
\details{
Description says all
}
\value{
Just the logarithm of the square of the argument
}
\author{
G P Nason
}

\seealso{\code{\link{image.wd}}, \code{\link{image.wst}}
}
\examples{
logabs(3)
# [1] 1.098612
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{math}
