\name{lt.to.name}
\alias{lt.to.name}
\title{
Convert notations (wavelet)
}
\description{
Convert level/notation into imwd.object component names
}
\usage{
lt.to.name(level, type)
}
\arguments{
\item{level}{
The level of the decomposition
}
\item{type}{
The type, one of CD, DC or DD, reflecting the application of filters
g,h in combination.
}}
\value{
A string in the imwd.object wavelet coefficient name format.
}
\details{
Not for user use!
}
% Converted by Sd2Rd version 0.3-3.
