\name{modernise.wd}
\alias{modernise.wd}
\title{Modernise a wd class object}
\usage{
\method{modernise}{wd}(wd, ...)
}
\arguments{
\item{wd}{The wd object you wish to modernise}
\item{...}{Other arguments}
}
\description{
Upgrade a version 2 \code{\link{wd.object}} to version 4. 
The function \code{\link{IsEarly}} can tell if the object comes from
an earlier version of WaveThresh.
}
\details{
Description says all.
}
\value{
The modernised object.
}
\author{G P Nason}
\keyword{manip}
