\name{newsure}
\alias{newsure}
\usage{
newsure(s, x)
}
\arguments{
\item{s}{Vector of standard deviations of coefficients}
\item{x}{Vector of regular (ie non-normalized) coefficients}
}
\title{Version of sure that acts as subsidiary for threshold.irregwd}
\description{
Version of the \code{\link{sure}} function used as a subsidiary for
\code{\link{threshold.irregwd}}.
}
\details{
Description says all
}
\value{
The SURE threshold 
}
\seealso{\code{\link{sure}}, \code{\link{threshold.irregwd}}}
\author{Arne Kovac}
\keyword{math}
