\name{print.BP}
\alias{print.BP}
\title{Print top best basis information for BP class object}
\usage{
\method{print}{BP}(x, \dots)
}
\arguments{
\item{x}{The BP object you wish to print}
\item{\dots}{Other arguments}
}
\description{
The function \code{\link{Best1DCols}} works out what are the best
packets in a selection of packets. This function prints out what the
best packet are.

The \code{\link{Best1DCols}} is not intended for user use, and hence neither
is this print method.
}
\details{
Description says all
}
\value{
None.
}
\seealso{\code{\link{Best1DCols}}}
\author{G P Nason}
\keyword{print}
