\name{print.imwdc}
\alias{print.imwdc}
\title{Print an `imwdc' Object}
\usage{
print.imwdc(x)
}
\description{
Prints out information about an imwdc object and then calls
\code{summary()} on it.
}
\arguments{
  \item{x}{an object of class \code{imwdc}}}
\section{Side Effects}{
Information about the \code{imwdc} object is printed out.
}
\seealso{\code{\link{imwdc.object}}.
}
% Converted by Sd2Rd version 0.3-3.
