\name{print.w2d}
\alias{print.w2d}
\title{Print method for printing w2d class objects}
\usage{
\method{print}{w2d}(x, \dots)
}
\arguments{
\item{x}{The w2d class object that you wish to print info about}
\item{\dots}{Other arguments}
}
\description{
Prints information about a w2d class object. These objects are not typically
directly used by a user.
}
\details{
Description says all
}
\seealso{\code{\link{wpst2discr}}}
\value{
Nothing
}
\author{G P Nason}
\keyword{print}
