\name{print.wd}
\alias{print.wd}
\title{Print Method for a `wd' object.}
\description{
  Prints out information about an wd object and then calls \code{summary()}
  on it.
}
\usage{
\method{print}{wd}(x, \dots)
}
\arguments{
  \item{x}{an object of class \code{wd}}
  \item{\dots}{further arguments to be passed to or from methods.}
}
\section{Side Effects}{
  Information about the \code{wd} object is printed out.
}
\seealso{\code{\link{wd}}, \code{\link{wd.object}}.
}
\keyword{utilities}

