\name{print.wpstRO}
\alias{print.wpstRO}
\title{
Print a wpstRO class object
}
\description{
Prints out a representation of an wpstRO object
}
\usage{
\method{print}{wpstRO}(x, maxbasis = 10, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
	The wpstRO object to print
}
  \item{maxbasis}{
	The maximum number of basis packets to report on
}
  \item{\dots}{
	Other arguments
}
}
\details{
Description says all
}
\value{
None
}
\author{ G P Nason }

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{\code{\link{makewpstRO}}
}
\examples{
#
# See example in makewpstRO function
#
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{print}
