\name{teddy}
\docType{data}
\alias{teddy}
\title{Picture of a teddy bear's picnic.}
\description{
A 512x512 matrix. Each entry of the matrix contains an image intensity value.
}
\usage{
data(teddy)
}
\format{
A 512x512 matrix. Each entry of the matrix contains an image intensity value. The whole matrix represents an image of a teddy bear's picnic.
}
\source{
Taken by Guy Nason. 
}
\examples{
#
# This command produces the image seen above.
#
# image(teddy)
#
}
\keyword{datasets}
\author{G P Nason}
