\name{threshold}
\alias{threshold}
\title{Thresholding Wavelets - Generic function}
\description{Generic function for soft or hard thresholding of wavelets.
}
\usage{
threshold(x, \dots)
}
\arguments{
  \item{x}{an \R object.}
  \item{\dots}{methods may have additional arguments.}
}
\value{
thresholding is applied to the object and a thresholded version is
returned.
}
\references{see \code{\link{wd}} for a list.
}
\seealso{
\code{\link{threshold.wd}}, \code{\link{threshold.imwd}} and
\code{\link{threshold.imwdc}} for the wavelet methods and \bold{examples}.

\code{\link{wd}}, \code{\link{wr}}, \code{\link{imwd}}, etc.
}
\keyword{smooth}
\keyword{nonlinear}
% Converted by Sd2Rd version 0.3-3.
