\name{uncompressed.object}
\alias{uncompressed.object}
\title{Uncompressed (Wavelet) Object}
\description{
  These are objects of class \code{"uncompressed"}.
  They are lists with components representing a vector, see below.

  Typically, these are generated by \code{\link{compress.default}}.
}
\value{
  The following components must be included in a legitimate
  \code{uncompressed} object.
  \item{vector}{the original uncompressed vector}
}
\seealso{\code{\link{compressed.object}} for details; \code{\link{compress}}.
}
\keyword{classes}
% Converted by Sd2Rd version 0.3-3.
