\name{wavethresh-package}
\alias{wavethresh-package}
\alias{wavethresh}
\docType{package}
\title{\packageTitle{wavethresh}}
\description{\packageDescription{wavethresh}}

\details{
\packageDESCRIPTION{wavethresh}
\packageIndices{wavethresh}
See book or individual help pages for main functions. For example,
\code{\link{wd}} for the one-dimensional discrete wavelet transform.
}
\author{
\packageAuthor{wavethresh}

Maintainer: \packageMaintainer{wavethresh}
}
\references{
Nason, G.P. (2008) Wavelet methods in Statistics with R. Springer, New York.
\href{https://link.springer.com/book/10.1007/978-0-387-75961-6}{Book URL.}
}
\keyword{math}
\seealso{\code{\link{ewspec}}, \code{\link{imwd}}, \code{\link{threshold}}, \code{\link{wd}}, \code{\link{wst}}
}
\examples{
#
# See examples in individual help pages
#
}
