\name{Whistory}
\alias{Whistory}
\usage{
Whistory(\dots)
}
\arguments{
\item{\dots}{Arguments to pass to method}
}
\title{Obsolete function supposedly detailed history of object}
\description{
The original idea behind this obsolete function was to interrogate
an object and return the modifications that had been successively
applied to the function. The reason for this was that after a long
data analysis session one would end up with a whole set of, e.g.,
thresholded or otherwise modified objects and it would have been convenient
for each object not only to store its current value but also the history
of how it got to be that value.
}
\details{
Description says all
}
\value{
No return value, although function was meant to print out a list
times and dates when the object was modified.
}
\seealso{\code{\link{Whistory.wst}}}
\author{G P Nason}
\keyword{utilities}
