\name{Whistory.wst}
\alias{Whistory.wst}
\title{Obsolete function: as Whistory, but for wst objects}
\usage{
\method{Whistory}{wst}(wst, all=FALSE, \dots)
}
\arguments{
\item{wst}{The object that you want to display the history for}
\item{all}{Print the whole history list}
\item{\dots}{Other arguments}
}
\description{
Obsolete function, see \code{\link{Whistory}}.
}
\details{
Description says all
}
\value{
Nothing, but history information is printed.
}
\seealso{\code{\link{Whistory}}}
\author{G P Nason}
\keyword{utilities}
