\name{wr.mwd}
\alias{wr.mwd}
\title{Multiple wavelet reconstruction for mwd objects}
\usage{
\method{wr}{mwd}(...)
}
\arguments{
\item{\dots}{Arguments to the \code{\link{mwr}} function.}
}
\description{
This function is method for the \code{\link{function}} to
apply the inverse multiple wavelet transform for \code{\link{mwd.object}}
objects. 
}
\details{
The function is merely a wrapper for \code{\link{mwr}}
}
\value{
The same return value as for \code{\link{mwr}}.
}
\seealso{
\code{\link{mwd}},
\code{\link{mwr}}}
\author{Tim Downie}
\keyword{math}
