\name{wr3D}
\alias{wr3D}
\title{Inverse DWT for 3D DWT object. }
\description{
Performs the inverse DWT for \code{\link{wd3D.object}}, i.e. 3D DWT objects. 
}
\usage{
wr3D(obj)
}
\arguments{
\item{obj}{A \code{\link{wd3D.object}} 3D DWT object as returned by \code{\link{wd3D}}.  }
}
\details{
The code implements a 3D version of Mallat's inverse pyramid algorithm. 
}
\value{
A 3D array containing the inverse 3D DWT of obj. 
}
\section{RELEASE}{Version 3.9.6 Copyright Guy Nason 1997}
\seealso{
\code{\link{wr}}, \code{\link{accessD.wd3D}}, \code{\link{print.wd3D}}, \code{\link{putD.wd3D}}, \code{\link{putDwd3Dcheck}}, \code{\link{summary.wd3D}}, \code{\link{threshold.wd3D}}, \code{\link{wd3D}}, \code{\link{wd3D.object}}. 
}
\examples{
#
# Now let's take the object generated by the last stage in the EXAMPLES
# section of threshold.wd3D and invert it!
#
#testwr <- wr3D(testwd3DT)
#
# You'll find that testwr is an array of dimension 8x8x8!
#
}
\keyword{smooth}
\keyword{nonlinear}
\author{G P Nason}
