\name{wvrelease}
\alias{wvrelease}
\title{Prints out the release number of the WaveThresh package}
\usage{
wvrelease()
}
\arguments{
None.}
\description{
PRints out the release number of the WaveThresh package, and some
copyright info.
}
\details{
Description says all
}
\value{
Nothing
}
\examples{
wvrelease()

}
\author{G P Nason}
\keyword{misc}
