\name{compress}
\alias{compress}
\title{
Compression - Generic function (wavelet)
}
\description{
Compresses object by removing zeroes
This is a generic function.
Functions with names beginning in \code{compress.} will be
methods for this function.
\code{imwd}, \code{default}
}
\usage{
compress(x, \dots)
}
\arguments{
\item{x}{an \R object.}
\item{...}{methods may have additional arguments.}
}
\value{
a compressed version of the supplied argument
}
\details{
See individual help pages for operation
}
\section{RELEASE}{
Release 2.2
Copyright Guy Nason 1993
}
\seealso{
\code{\link{threshold}}, \code{\link{uncompress}}
}
% Converted by Sd2Rd version 0.3-3.
