\name{imwr}
\alias{imwr}
\title{2D Inverse Discrete Wavelet Transform (Image W. Reconstruction)}
\description{
  Generic function for 2-dimensional wavelet reconstruction.
}
\usage{
imwr(imwd, \dots)
}
\arguments{
  \item{imwd}{a wavelet decomposition object.}
  \item{\dots}{methods may have additional arguments.}
}
\seealso{the principal method, \code{\link{imwr.imwd}},
  and \code{\link{imwr.imwdc}}.
}
\keyword{nonlinear}
\keyword{smooth}
