\name{lt.to.name}
\alias{lt.to.name}
\title{Convert notations (wavelet)}
\description{
  Convert level/notation into \code{imwd.object} component names.

  Not for user use!
}
\usage{
lt.to.name(level, type)
}
\arguments{
  \item{level}{the level of the decomposition.}
  \item{type}{the type, one of CD, DC or DD, reflecting the application
    of filters g,h in combination.}
}
\value{
A string in the imwd.object wavelet coefficient name format.
}
\keyword{utilities}
% Converted by Sd2Rd version 0.3-3.
