\name{print.imwd}
\alias{print.imwd}
\alias{print.imwdc}
\title{Print an `imwd' or `imwdc' object}
\description{
  Prints out information about the object and then calls \code{summary()}
  on it.
}
\usage{
print.imwd(x, \dots)
print.imwdc(x, \dots)
}
\arguments{
  \item{x}{an object of class \code{imwd} or \code{imwdc}, respectively}.
}
\details{
  Contributed in release 2.2 by Martin Maechler
}
\seealso{\code{\link{imwd}}, \code{\link{imwd.object}},
  \code{\link{imwdc.object}.
}
\keyword{utilities}
% Converted by Sd2Rd version 0.3-3.
