\name{print.imwd}
\alias{print.imwd}
\title{Print an `imwd' object}
\description{
Prints out information about an imwd object and then calls \code{summary()}
on it.
}
\usage{
print.imwd(x)
}
\arguments{
\item{x}{an object of class \code{imwd}}.}
\details{
Contributed in release 2.2 by Martin Maechler
}
\section{Side Effects}{
Information about the \code{imwd} object is printed out.
}
\seealso{\code{\link{imwd}}, \code{\link{imwd.object}}.
}
% Converted by Sd2Rd version 0.3-3.
