\name{uncompress}
\alias{uncompress}
\title{
Decompression - Generic function (wavelet)
}
\description{
Decompresses object by padding with zeroes
This is a generic function.
Functions with names beginning in \code{uncompress.} will be
methods for this function.
\code{imwdc}, \code{default}
}
\usage{
uncompress(x, \dots)
}
\arguments{
  \item{x}{an \R object.}
  \item{\dots}{methods may have additional arguments.}
}
\value{
a uncompressed version of the supplied argument
}
\details{
See individual help pages for operation
}
\seealso{
\code{\link{compress}}
}
% Converted by Sd2Rd version 0.3-3.
