\name{wavethresh-package}
\alias{wavthresh-package}
\alias{wavethresh}
\docType{package}
\title{Wavelet transforms and associated statistical methodology.
}
\description{Performs one-, two- and three-dimensional wavelet
	transforms, nondecimated transforms, wavelet packet transforms,
	nondecimated wavelet packet transforms,
	complex-valued wavelet transforms, wavelet shrinkage for various
	kinds of data, locally stationary wavelet time series,
	nonstationary multiscale transfer function modeling,
	density estimation.
}
\details{
\tabular{ll}{
Package: \tab wavethresh \cr
Type: \tab Package\cr
Version: \tab 4.6.6\cr
Date: \tab 2013-10-21\cr
License: \tab GPL (>=2)\cr
}
}
\author{Guy Nason, <g.p.nason@bristol.ac.uk>
}
\references{
Nason, G.P. (2008) Wavelet methods in Statistics with R. Springer, New York.
\href{http://www.springer.com/statistics/statistical+theory+and+methods/book/978-0-387-75960-9}{Book URL.}

Errata and fast-breaking news: \url{http://www.stats.bris.ac.uk/~wavethresh}

}
\keyword{math}
\seealso{\code{\link{ewspec}}, \code{\link{imwd}}, \code{\link{threshold}}, \code{\link{wd}}, \code{\link{wst}}
}
\examples{
#
# See examples in individual help pages
#
}
